frontend(rights): move rights as a sub-folder in organization lib
