+++ libs/organization/src/index.ts
@@ -2,3 +2,5 @@ export * from './lib/+state';
 export * from './lib/organization.module';
 export * from './lib/guards/organization-list.guard';
 export * from './lib/org-form/org-form.component';
+export * from './lib/rights/guard/rights.guard'
+export * from './lib/rights/+state/rights.service'
