+++ libs/organization/src/lib/rights/+state/rights.model.ts
