+++ libs/organization/src/lib/rights/index.ts
@@ -1,0 +1,2 @@
+export * from './+state/rights.model';
+export * from './+state/rights.service';
