--- libs/rights/README.md
@@ -1,7 +1,0 @@
-# libs-rights
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `ng test libs-rights` to execute the unit tests.
