--- libs/rights/src/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/rights.module';
-export * from './lib/guards/rights.guard';
-export * from './lib/+state/rights.model';
-export * from './lib/+state/rights.service';
