--- libs/rights/src/lib/rights.module.ts
@@ -1,5 +1,0 @@
-import { NgModule } from '@angular/core';
-
-@NgModule({
-})
-export class RightsModule {}
