--- libs/rights/src/test-setup.ts
@@ -1,1 +1,0 @@
-import 'jest-preset-angular';
