--- libs/rights/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "types": ["jest", "node"]
-  }
-}
