--- libs/rights/tslint.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../tslint.json",
-  "rules": {
-    "directive-selector": [true, "attribute", "blockframes", "camelCase"],
-    "component-selector": [true, "element", "blockframes", "kebab-case"]
-  }
-}
