+++ tsconfig.json
@@ -25,8 +25,7 @@
       "@blockframes/account": ["libs/account/src/index.ts"],
       "@blockframes/material": ["libs/material/src/index.ts"],
       "@blockframes/template": ["libs/material/src/index.ts"],
-      "@blockframes/contracts": ["contracts/build"],
-      "@blockframes/rights": ["libs/rights/src/index.ts"]
+      "@blockframes/contracts": ["contracts/build"]
     },
     "module": "es2015",
     "rootDir": "."
